move wave-button
