+++ shared/chat/blocking/invitation-to-block.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react'
 import * as Kb from '../../common-adapters'
-import {WaveButton} from '../../settings/contacts-joined/buttons'
 import * as Styles from '../../styles'
 import * as Container from '../../util/container'
 import * as ProfileGen from '../../actions/profile-gen'
@@ -35,7 +34,7 @@ const BlockButtons = (props: Props) => {
       style={styles.button}
     >
       {!team && (
-        <WaveButton small={true} usernames={[adder, ...(others || [])].join(',')} style={styles.button} />
+        <Kb.WaveButton small={true} usernames={[adder, ...(others || [])].join(',')} style={styles.button} />
       )}
       {!team && others.length === 0 && (
         <Kb.Button
